% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{A heuristic solution to the knapsack problem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe.}

\item{W}{A non-negative number.}
}
\value{
A list detailing the maximum value and the elements to get it.
}
\description{
A heuristic solution to the knapsack problem
}
