\name{khwarizmi-package}
\alias{khwarizmi-package}
\alias{khwarizmi}
\docType{package}
\title{
\packageTitle{khwarizmi}
}
\description{
\packageDescription{khwarizmi}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{khwarizmi}

Maintainer: \packageMaintainer{khwarizmi}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = khwarizmi)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{khwarizmi}}
%\section{Documentation Index}{\packageIndices{khwarizmi}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
