% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Solving the knapsack problem in the case of discrete weights}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A dataframe.}

\item{W}{A non-negative number.}
}
\value{
A list detailing the maximum value and the elements to get it.
}
\description{
Solving the knapsack problem in the case of discrete weights
}
