% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force the knapsack problem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A dataframe.}

\item{W}{A non-negative number.}
}
\value{
A list detailing the maximum value and the elements to get it.
}
\description{
Brute force the knapsack problem
}
